refactor(vscode): share initialization options
