+++ editors/vscode/tsconfig.browser.json
@@ -7,10 +7,10 @@
   "include": [
     "../../packages/vide-extension-shared/src/**/*.ts",
     "src/browser/**/*.ts",
+    "src/common/**/*.ts",
     "src/diagnosticActions.ts",
     "src/diagnosticRules.ts",
     "src/generated/**/*.ts",
-    "src/initializationOptions.ts",
     "src/projectConfigCommon.ts",
     "src/status.ts",
     "src/videStatus.ts"
